chore(rename): Add Pv prefix for system app components
