+++ apps/client/src/widgets/common/ui/PvApp.vue
