+++ apps/client/src/widgets/common/ui/PvAppMain.vue
